Fix printf type mismatches in debug output
Updates logging/print statements across board, DDP, serial, L6470 JSON, and remote config code to use explicit unsigned casts with `%u`, avoiding signed/width mismatches. Also adds a status LED mode debug print, cleans up minor formatting in `modeparams.cpp`, renames an intermediate float variable for clarity in motor params, and updates the modeparams copyright year range.

+++ lib-board/src/h3/board_statusled.cpp
@@ -59,11 +59,11 @@ void SetFrequency(uint32_t frequency_hz) {
 
     if (s_timer_id < 0) {
         s_timer_id = SoftwareTimerAdd((1000U / frequency_hz), Ledblink);
-        DEBUG_PRINTF("s_timer_id=%d", s_timer_id);
+        DEBUG_PRINTF("s_timer_id=%u", static_cast<unsigned>(s_timer_id));
         return;
     }
 
-    DEBUG_PRINTF("s_timer_id=%d", s_timer_id);
+    DEBUG_PRINTF("s_timer_id=%u", static_cast<unsigned>(s_timer_id));
     SoftwareTimerChange(s_timer_id, (1000U / frequency_hz));
 }
 } // namespace board::statusled
